Sync with NetBSD: Allocate SCSI message buffers in ncr53c9x_attach(), with a provision for front-ends to override the allocation to avoid alignment handling in their DMA engines. Note that that ncr53c9x_msgout() can request a 1 byte DMA transfer that would be difficult to break up.
Eliminate the degenerate loop in `ncr53c9x_intr()'. Retain the delay heuristic
it implemented under the label `shortcut:' and only use it in these cases:
  (1) after successful re-relection,
  (2) after receiving command-complete status, and
  (3) during message-in handshake.

If we have an NCR53CF9x (indicated by the front-end by setting a flag
in the softc's new "features" word), set FSCSI bit in CFG3 if the sync
period is <= 200ns, or clear it otherwise.