Change extmem counter to include *all* extended memory segments.
Obviously, the kernel does not (and can not) know where the ACPI and such
pages are, since the boot -> kernel API does not support passing the
information.

This will stomp on them, once again, but we can NOT limit half the PCs
in existance to only 64MB of RAM.

Approved by deraadt.