From NetBSD (bouyer@):
In siop_reset(), reset sc_ntargets to 0. The number of targets will be
computed again in siop_add_reselsw().

In siop_reset(), reset the tag reseloff to 0, in addition to the lun
reseloff. If siop_add_dev() fails this time we would use the old
reseloff, clobbering memory now used for something else.