Add support for all six Motorola-supported configurations of MVME376 cards. The original code would only work with the first configuration, as the etherbuf addresses grow _downwards_ from 0xfd6c0000 with Motorola's jumper configuration.
Reference: MVME376 Ethernet Communication Controller User's Manual (MVME376/D1)

While there, do not panic() on failure to attach, but simply print an
error message and return.

XXX Proper care of the memory mapping still needs to be written, as the memory
XXX space of the MVME376 is always outside extio.