The device doesn't always detach gracefully from the bus after a firmware upload. We thus force a port reset and a re-exploration on the parent hub after firmware upload.
In uath_newstate(), if we're called from a process context, we don't need
to schedule a USB task. This fixes a problem in uath_stop(): when moving
to the S_INIT state, the task was scheduled after the device was reset.

Don't abort the RX firmware command pipe in uath_stop(), otherwise we will
never receive commands ACKs from the firmware anymore.  We now survive to
an ifconfig down up.